EnzoNG wrapper: C-ABI shim + HDF5-free raw writer for streaming ICs
Wrap the makeCosICs realizer as a CodeBridge C-ABI library for the
EnzoNG.jl multi-code framework, so a Julia-driven session can generate
baryon–dark-matter streaming-velocity ICs in-process and inject them
into live Enzo and RAMSES.

- makeCosICs/capi_out.c: WriteCapiSnapshot — HDF5-free `.cicass` raw dump
  of {DM particles + gas grid + streaming offset} (magic CICASS01); dodges
  the libhdf5/libenzo collision.
- makeCosICs/cicass_capi.cc: C-ABI shim (cicass_generate/last_error/
  version/real_bytes).
- makeCosICs/main.c: dual-purpose entry (CICASS_LIB → cicass_run_genics,
  else main) + the -DOUTPUT_CAPI branch.
- deps/build_cicass_darwin.sh: builds transfer.x (GSL) and
  libcicass_capi.dylib (FFTW3+GSL, no HDF5) via Homebrew.
- vbc_transfer/main.cc: modern-clang (size_t)NUMEQNS narrowing fix.

+++ vbc_transfer/main.cc
@@ -402,7 +402,7 @@ void returnGrowth(double G[], double z, double k, double costh)
 
 #ifdef GSL
   int calcGrowthIntGSL(double a, const double *y, double *dy, void *params);
-  gsl_odeiv2_system sys = {calcGrowthIntGSL, NULL, NUMEQNS, NULL};
+  gsl_odeiv2_system sys = {calcGrowthIntGSL, NULL, (size_t)NUMEQNS, NULL};
 
   gsl_odeiv2_driver * d =  gsl_odeiv2_driver_alloc_y_new(&sys, gsl_odeiv2_step_rkf45,
 				    1e-3, 1e-6, 0.0);
